# backend/core/generated/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml

from __future__ import annotations

from datetime import date as date_aliased
from enum import Enum

from pydantic import BaseModel, ConfigDict, EmailStr, Field, RootModel, constr


class Timezone(Enum):
    Europe_Moscow = "Europe/Moscow"


class ErrorDetail(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    message: str
    fieldErrors: dict[str, list[str]] | None = None


class PageSize(Enum):
    number_20 = 20


class Code(Enum):
    INVALID_REQUEST = "INVALID_REQUEST"


class InvalidRequestBody(ErrorDetail):
    model_config = ConfigDict(
        extra="forbid",
    )
    code: Code


class MoscowDate(RootModel[date_aliased]):
    root: date_aliased = Field(..., description="Moscow calendar date (YYYY-MM-DD).")


class MoscowInstant(
    RootModel[constr(pattern=r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?\+03:00$")]
):
    root: constr(pattern=r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?\+03:00$") = Field(
        ..., description="ISO 8601 instant with the fixed Moscow offset."
    )


class Status(Enum):
    available = "available"
    busy = "busy"


class Slot(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    startAt: MoscowInstant
    endAt: MoscowInstant
    status: Status


class Code1(Enum):
    SLOT_UNAVAILABLE = "SLOT_UNAVAILABLE"


class SlotUnavailableBody(ErrorDetail):
    model_config = ConfigDict(
        extra="forbid",
    )
    code: Code1


class AvailabilityDate(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    date: MoscowDate
    slots: list[Slot]


class BookingRequest(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    startAt: MoscowInstant
    name: constr(min_length=1, max_length=255) = Field(
        ..., description="Trimmed before validation; whitespace-only names are invalid."
    )
    email: EmailStr


class Event(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    id: str
    name: str
    email: str
    startAt: MoscowInstant
    endAt: MoscowInstant
    createdAt: MoscowInstant


class EventsPage(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    items: list[Event]
    page: int
    pageSize: PageSize
    total: int
    nextPage: int | None = Field(...)


class Availability(BaseModel):
    model_config = ConfigDict(
        extra="forbid",
    )
    timezone: Timezone
    dates: list[AvailabilityDate] = Field(
        ...,
        description="Exactly 14 dates, from today in Moscow through the following 13 days.",
        max_length=14,
        min_length=14,
    )